New tool reorg_spamdb:  reorg_spamdb copies the content into a fresh hash database, so the size will be hopefully decreased.  Owner of new db is every time _spamd:_spamd, access mode will be the same as from from source db.  With parameter -r the tool replace the old db with the fresh one (only if source is in hash format).
 Thanks to Robert Schulze rs @ bytecamp.net, for suggesting such a tool.

spamdb.c:
 add parameter -W (whiteexp)
 casts to make compiler at amd64/sparc happy
 rewrite parts of spamdb(8) man page
 
spamd.c:
 enforce user _spamd
 add tmp variables to for passtime,greyexp,whiteexp to make compiler at amd64/sparc happy

sync.c
 cast for amd64/sparc

+++ trunk/spamd/sync.c
@@ -356,7 +356,7 @@ sync_recv(void)
 				/* send this info to the greylister */
 				fprintf(grey, "WHITE:%s:", inet_ntoa(ip));
 				fprintf(grey, "%s:%u\n",
-				    inet_ntoa(addr.sin_addr), expire);
+				    inet_ntoa(addr.sin_addr), (u_int32_t)expire);
 				fflush(grey);
 			}
 			break;
@@ -377,7 +377,7 @@ sync_recv(void)
 				/* send this info to the greylister */
 				fprintf(grey, "TRAP:%s:", inet_ntoa(ip));
 				fprintf(grey, "%s:%u\n",
-				    inet_ntoa(addr.sin_addr), expire);
+				    inet_ntoa(addr.sin_addr), (u_int32_t)expire);
 				fflush(grey);
 			}
 			break;
